minor refactor added todo
