+++ management/server/core/subutai-appender/pom.xml
@@ -16,10 +16,6 @@
 
   <dependencies>
     <dependency>
-      <groupId>io.subutai</groupId>
-      <artifactId>subutai-hub-share</artifactId>
-    </dependency>
-    <dependency>
       <groupId>log4j</groupId>
       <artifactId>log4j</artifactId>
       <version>${log4j.version}</version>
