+++ management/server/subutai-common/src/main/java/io/subutai/common/peer/Peer.java
@@ -278,8 +278,10 @@ public interface Peer extends RelationLink
 
     void addReverseProxy( ReverseProxyConfig reverseProxyConfig ) throws PeerException;
 
+    @Deprecated
     void addCustomProxy( CustomProxyConfig proxyConfig ) throws PeerException;
 
+    @Deprecated
     void removeCustomProxy( CustomProxyConfig proxyConfig ) throws PeerException;
 
     SshKeys getSshKeys( EnvironmentId environmentId, SshEncryptionType sshEncryptionType ) throws PeerException;
