+++ management/server/subutai-common/src/main/java/io/subutai/common/protocol/ReverseProxyConfig.java
@@ -8,6 +8,8 @@ import io.subutai.common.network.ProxyLoadBalanceStrategy;
 
 /**
  * DTO object for reverse proxy config
+ *
+ * TODO add port
  */
 public class ReverseProxyConfig
 {
